Project Organizations
Going to try making a proper app out of this concept, been wanting to do that for a long time now! Just made this repo and committed the original history to it tonight, this is the real date of when this repo was made essentially. The project itself goes back to the commit dates though, I like tracking things properly with their dates in that way. There are a few legacy versions of Num-Text and Menu-Drop that I wished I had committed back when I made their repos too, from the *very* early-on days. Those still haven't been catalogged in their repos yet.

I love the guitar for 'No King No Princess'. Wow these guys are catchy. I already knew that before, but even moreso now too! Loving it!! This is the stage of the year where I find new music I think. Seems like summer and spring is for writing, fall and winter are for listening. This seemed to happen last year, and I didn't know of this pattern explicitly yet. Going to learn more about that.