Separation of Concerns Separations
Breaking apart the implementation of the app into it's own source files, rather than writing everything in a single file. I really loved the idea early on in my coding to be able to share projects directly, without needing to zip things, build things, compile things. It just opens, and you have your app. I see why I got into web dev, and I think that same kind of magic can still be applied, when you configure your tools to do the heavy lifting for you, rather than with you. I think that was my opposition to using tooling for a long time, it was that I saw it as complexity that didn't actually help with anything. Now I see it as a major help for productivity and maintainability, both of which were challenges I ended up facing with STE as it grew and aged as a codebase. Now I am STILL cleaning things up there, over all of the years of changing things from how they were implemented over time. It still needs a heck ton of work, mindblowing.