Vite! (and GitHub Pages) 🤘
Trainfire.

Didn't realize how interesting of a mix the themes are on Ki and DTP overall. I thought it was a solely calm record, but it has metal, the brush jazz parts on the train sections, very interesting! You can hear how that's a crossover with the other albums in the original 4. Deconstruction is also train-themed, somewhat. Lady Helen is really cool. The self-titled song sounds great too so far, just a few seconds into it currently as we speak! (huh?)