Instrument State Class
Deduping the logic for this here, to make things easier to declare and manage. Now I see how it can be useful to hide state to a property on the class, and only expose certain public methods, then you don't have to worry about when you change the underlying implementation if the class's public API will implicitly change too, you can choose what to be explicit with.