Proper Keys Map
Taking a note from TS's type declarations for elements, but in the literal sense of the runtime itself. I generate a lookup table that associates a given key with it's instrument object, and simply accessing the key itself depends on whether a keypress was associated with anything when it was defined.

You can log out the `keysMap` itself to see what it looks like in the literal sense. Works awesome!!! I think this is much more performant for static lookup too.