Web Audio API (v3!)
This is the next stage of the project, now it's built with the Web Audio API, which is much more performant! At least, it's much more responsive.

Now I want to try making it so that the app can run on a single AudioContext instance, I'm not sure how that will work out though, like for pausing the playback of the hi-hat related sounds accordingly.

Next on the agenda is to try adding a better GUI, and gamepad support! I have recently been curious abount merging this with my gamepad-observer project so you can write MIDI drum tracks with your game controller, seems like a cool concept, I haven't seen anyone do yet. The MIDI file part I only just put together last night, I haven't explicitly looked into that directly yet. It could work great with Logic and GarageBand though, and other tools in general.

I also fixed the mixing for some of the sounds in this update, made the start and end times a little tighter, and changed the volume for some of them. The hi-hat is louder overall, crash a little quieter, and the snare a little quieter.