Proper Gamepad Triggers!
WOW! It works!!! This is the first time I have it actually in the flesh! A crack in the ice, appeaaars underneath your feet.

The mappings could use work, I think the Instrument class could instead be moved over to a Solid component, and it's really late and I have to get up early (4:30 eek). See ya tomorrow. Very happy with the success here!