Safari Support
Baah, this is a mess! Safari didn't work at all, for multiple reasons.

Gamepad objects are saved, so their values are updated. Using this to diff your Solid signals doesn't work because it's the same object reference. Chrome returns a different object with each gamepad change.

Gamepad button pressed is only true when the button is *fully* pressed, rather than partially pressed. Chrome does the other way around, any press is a press.

At first I thought the audio wasn't playing because of audio playback support, and maybe it was just a load thing.
https://stackoverflow.com/questions/57510426/cannot-resume-audiocontext-in-safari (This doesn't appear to be the issue afterall)
It does appear to be a permissions issue though, and DevTools or the API call doesn't give any context as to why this isn't working. It doesn't error, no warnings in the console. So you have to request for audio access by a user gesture before you can play sound with an audio context. Makes sense, but why is it so difficult to find this? Kind of unrelated, but you can't use `navigator.mediaDevices.getUserMedia()` unless you're in a secure context, but that means debugging over the network on your mobile device didn't initially work, because you would need to use HTTPS to host it with, which is also a mess. Thankfully there is an option to configure allowing it in DevTools, yay! That was going to be close. Thankfully, now everything is debugged, and I can play drums with my controller on my iPad! Was trying it on my phone first, and none of this worked, so I had to take probably 45 minutes to debug it, I thought it was going to work, but nah. Not so fast!!! 🥹
https://stackoverflow.com/questions/44523609/getusermedia-is-not-allowed-in-localhost-safari-11

I'm not sure if I want to make a library over the top of the Gamepad API fully at this point, it really is kind of a mess to use, in terms of browser compatibility, and consistency. Might change gamepad-state to do that. We'll see. May not be as hard as I think it is. It doesn't make sense for me to build the nice project for that to abstract things away, but then still have to do it manually myself anyways. I think I'll have to take a second go-around at making it streamlined :)

https://github.com/Offroaders123/gamepad-state/issues/3